requirements.txt
pandas==2.1.1
huggingface-hub==0.17.3
Pillow==10.0.1
deepdanbooru==1.0.2
opencv-python==4.8.0.76
flask
requests



